add react-paginate  to properties page
