+++ package.json
@@ -22,6 +22,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.54.2",
+    "react-paginate": "^8.3.0",
     "react-router": "^7.2.0",
     "swiper": "^11.2.2",
     "yup": "^1.6.1"
